check for file existing

+++ compress.sh
@@ -8,5 +8,10 @@ do
   gltf-pipeline -i "$filename".glb -o "$filename".glb -d -b
   b3dm pack -i "$filename".glb -f
   rm "$filename".glb
-  rm "$filename".batch
+  if [ -f "$filename".batch ]; then
+    rm "$filename".batch
+  fi
+  if [ -f "$filename".featuretable.json ]; then
+    rm "$filename".featuretable.json
+  fi
 done
